feat: Insert the table orders and projects

+++ LearningCenter.API/Security/Domain/Repositories/IUserRepository.cs
@@ -6,10 +6,10 @@ public interface IUserRepository
 {
     Task<IEnumerable<User>> ListAsync();
     Task AddAsync(User user);
-    Task<User> FindByIdAsync(int id);
+    Task<User> FindByIdAsync(long id);
     Task<User> FindByUsernameAsync(string username);
     bool ExistsByUsername(string username);
-    User FindById(int id);
+    User FindById(long id);
     void Update(User user);
     void Remove(User user);
 }
